refactor: Extract operator to enum
