+++ RacingCar/Calculator.swift
@@ -23,32 +23,15 @@ struct Calculator: Calculable {
 		return try calculate(by: expressions)
 	}
 	
-	// MARK: - Arithmetic Operation
-	private func add(_ lhs: Int, _ rhs: Int) -> Int {
-		return lhs + rhs
-	}
-	
-	private func subtract(_ lhs: Int, _ rhs: Int) -> Int {
-		return lhs - rhs
-	}
-	
-	private func multiply(_ lhs: Int, _ rhs: Int) -> Int {
-		return lhs * rhs
-	}
-	
-	private func divide(_ lhs: Int, _ rhs: Int) -> Int {
-		return lhs / rhs
-	}
-	
-	// MARK: - Others
+	// MARK: - Private
 	private func calculate(by expressions: [String]) throws -> Int {
 		var result = try expressions.first.toInt()
 		for i in findOperatorRange(in: expressions) {
 			guard expressions.count > i + 1 else { throw ValueError.invalid }
 			
 			let operand = try expressions[i + 1].toInt()
-			let arithmeticOperator = try find(opearator: expressions[i])
-			result = arithmeticOperator(result, operand)
+			let arithmeticOperator = findOperator(in: expressions[i])
+			result = try arithmeticOperator.operate(result, operand)
 		}
 		
 		return result
@@ -58,18 +41,7 @@ struct Calculator: Calculable {
 		return stride(from: 1, to: expressions.count, by: 2)
 	}
 	
-	private func find(opearator arithmeticOpearator: String) throws -> (Int, Int) -> Int {
-		switch(arithmeticOpearator) {
-		case "+":
-			return add
-		case "-":
-			return subtract
-		case "*":
-			return multiply
-		case "/":
-			return divide
-		default:
-			throw ValueError.invalidOperator
-		}
+	private func findOperator(in expressions: String) -> ArithmeticOpearator {
+		ArithmeticOpearator(rawValue: expressions) ?? .none
 	}
 }
